requirements.txt
# Core dependencies with compatible versions
crewai>=0.55.2,<0.80.0
crewai[tools]>=0.55.2,<0.80.0

# Pydantic compatibility fix
pydantic>=2.8.0,<3.0.0

# Deep learning models
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.7.0

# Vector search
faiss-cpu>=1.8.0

# Image processing
Pillow>=10.0.0
ultralytics>=8.2.0

# Machine learning
scikit-learn>=1.3.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Development
python-dotenv>=1.0.0
